rename MetamaskContext > MetaMaskContext
